event onResourceStart, other fixes and improvments
