+++ SlipeServer.Legacy/LegacyMtaServer.cs
@@ -2,41 +2,12 @@
 using Microsoft.Extensions.Logging;
 using SlipeServer.Server.Resources.Serving;
 using SlipeServer.ConfigurationProviders.Configurations;
-using Microsoft.Extensions.Options;
 using SlipeServer.Lua;
 using SlipeServer.Server.Resources.Interpreters;
 using Microsoft.Extensions.Logging.Console;
-using Microsoft.Extensions.Logging.Abstractions;
 using SlipeServer.Server.Resources;
 
 namespace SlipeServer.Legacy;
-
-public class CustomConsoleFormatter : ConsoleFormatter
-{
-    private readonly SimpleConsoleFormatterOptions formatterOptions;
-
-    public CustomConsoleFormatter(IOptionsMonitor<SimpleConsoleFormatterOptions> options)
-        : base("customFormatter")
-    {
-        this.formatterOptions = new SimpleConsoleFormatterOptions
-        {
-            TimestampFormat = "[HH:mm:ss] "
-        };
-    }
-
-    public override void Write<TState>(in LogEntry<TState> logEntry, IExternalScopeProvider? scopeProvider, TextWriter textWriter)
-    {
-        if (textWriter == null)
-        {
-            throw new ArgumentNullException(nameof(textWriter));
-        }
-
-        var timestamp = DateTime.Now.ToString(this.formatterOptions.TimestampFormat);
-        textWriter.Write(timestamp);
-        textWriter.Write(logEntry.Formatter(logEntry.State, logEntry.Exception));
-        textWriter.Write(Environment.NewLine);
-    }
-}
 
 public class LegacyMtaServer
 {
@@ -82,7 +53,6 @@ public class LegacyMtaServer
             services.AddSingleton<IResourceServer, BasicHttpServer>();
             services.AddHostedService<LegacyServerService>();
             services.AddHostedService<LegacyConsoleCommandsService>();
-            //services.AddHostedService<ResourcesServerHostedService>();
 
             services.TryAddSingleton<ILogger>(x => x.GetRequiredService<ILogger<MtaServer>>());
         });
