+++ SlipeServer.LegacyServer/mods/deathmatch/resources/hello/s.lua
@@ -1,8 +1,8 @@
 ﻿print("print hello world");
 outputDebugString("outputDebugString hello world");
 
-local eventAdded = addEventHandler("onResourceStart", root, function(startedResource, b)
-	print("lua started", startedResource, b)
+local eventAdded = addEventHandler("onResourceStart", root, function(startedResource)
+	print("lua started", getResourceName(startedResource))
 end)
 
 print("eventAdded", eventAdded)
